Move BufferProvider to Rx.NET.
Part of the point of this work is to create a buffer provider with an
IObservable overflow endpoint. Such a provider could implement
IObservable from scratch, but that seems foolish when this will normally
be used with Rx anyway. It is also possible that buffer providers will
be useful in other places in Rx, e.g. ObserveOn; they are unlikely to be
useful elsewhere in Ix due to the pull nature of the Enumerables.

However, if BufferProvider lives in Ix but uses Rx then we have an
assembly dependency, which is undesirable. Similarly, if Ix uses
IBufferProvider defined by Rx then we have a dependency the other way.
So change ToAsyncObservable to use a constructor function rather than a
factory object; then we can move BufferProvider into Rx.

I would like to provide an implicit cast from IBufferProvider to the
constructor function, but operators on interfaces require C#11.